resort-api/requirements.txt
requests==2.32.5
fastapi==0.124.2
mypy-boto3-dynamodb==1.42.3
mypy-boto3-s3==1.42.3
mangum==0.19.0